# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Health check",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered, tokens returned"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated, tokens returned"
          },
          "403": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login with email & password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New token pair issued (rotation)"
          },
          "403": {
            "description": "Invalid or expired refresh token"
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie-auth": []
          }
        ],
        "summary": "Refresh access token using refresh token (cookie or body)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Logout — invalidates refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/forms": {
      "post": {
        "operationId": "FormsController_createForm",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Form created, returns formId and currentStage"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Start a new multi-stage form",
        "tags": [
          "Forms"
        ]
      },
      "get": {
        "operationId": "FormsController_listForms",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "in-progress",
                "completed"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of forms with meta"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "List all forms for the authenticated user (paginated)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/pending": {
      "get": {
        "operationId": "FormsController_getPending",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of pending forms"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get all in-progress forms for the authenticated user",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}": {
      "get": {
        "operationId": "FormsController_getForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Form data with currentStage for resume"
          },
          "404": {
            "description": "Form not found"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get form by ID — use to resume from last saved stage",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/stage/1": {
      "post": {
        "operationId": "FormsController_saveStage1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Stage1Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Stage 1 — Basic Information (multipart/form-data)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/stage/2": {
      "post": {
        "operationId": "FormsController_saveStage2",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Stage2Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Stage 2 — Address Details (multipart/form-data)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/stage/3": {
      "post": {
        "operationId": "FormsController_saveStage3",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Stage3Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Stage 3 — Professional Details (multipart/form-data)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/stage/4": {
      "post": {
        "description": "Upload photoId (required), resume (required), and up to 3 additionalDocuments",
        "operationId": "FormsController_saveStage4",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Stage 4 — Document Upload (multipart/form-data)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/stage/5": {
      "post": {
        "operationId": "FormsController_saveStage5",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Stage5Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Stage 5 — Emergency Contact (multipart/form-data)",
        "tags": [
          "Forms"
        ]
      }
    },
    "/api/v1/forms/{id}/submit": {
      "post": {
        "operationId": "FormsController_submitForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Form submitted successfully"
          },
          "400": {
            "description": "Missing stage data — lists which stages are incomplete"
          }
        },
        "security": [
          {
            "cookie-auth": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Submit the form — validates all stages are complete and marks as submitted",
        "tags": [
          "Forms"
        ]
      }
    }
  },
  "info": {
    "title": "CX Assessment API",
    "description": "Enterprise-grade REST API with JWT authentication and multi-stage form workflow",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication & authorization endpoints"
    },
    {
      "name": "Forms",
      "description": "Multi-stage application form endpoints"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie-auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssword1!",
            "description": "Min 8 chars, at least 1 uppercase, 1 lowercase, 1 number, 1 special char"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssword1!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Stage1Dto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+919876543210",
            "description": "Phone number with country code"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1990-01-15"
          },
          "gender": {
            "type": "string",
            "example": "male",
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone"
        ]
      },
      "Stage2Dto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "example": "123 Main Street, Apt 4B"
          },
          "addressLine2": {
            "type": "string",
            "example": "Near Central Park"
          },
          "city": {
            "type": "string",
            "example": "Bangalore"
          },
          "state": {
            "type": "string",
            "example": "Karnataka"
          },
          "pincode": {
            "type": "string",
            "example": "560001"
          },
          "country": {
            "type": "string",
            "example": "India"
          }
        },
        "required": [
          "addressLine1",
          "city",
          "state",
          "pincode"
        ]
      },
      "Stage3Dto": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string",
            "example": "Acme Corp"
          },
          "designation": {
            "type": "string",
            "example": "Senior Software Engineer"
          },
          "yearsOfExperience": {
            "type": "string",
            "example": "5",
            "description": "Years of professional experience"
          },
          "skills": {
            "type": "string",
            "example": "TypeScript, NestJS, PostgreSQL, React",
            "description": "Comma-separated list of skills"
          },
          "linkedInUrl": {
            "type": "string",
            "example": "https://linkedin.com/in/johndoe"
          },
          "portfolioUrl": {
            "type": "string",
            "example": "https://johndoe.dev"
          }
        },
        "required": [
          "company",
          "designation",
          "yearsOfExperience"
        ]
      },
      "Stage5Dto": {
        "type": "object",
        "properties": {
          "emergencyContactName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "emergencyContactPhone": {
            "type": "string",
            "example": "+919876543210"
          },
          "emergencyContactRelationship": {
            "type": "string",
            "example": "spouse",
            "enum": [
              "spouse",
              "parent",
              "sibling",
              "friend",
              "colleague",
              "other"
            ]
          }
        },
        "required": [
          "emergencyContactName",
          "emergencyContactPhone",
          "emergencyContactRelationship"
        ]
      }
    }
  }
}